HTBHF-272 move formatters into directory
