+++ src/web/routes/application/common/formatters/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require('./dates')
+}
